server/service_cap_df/requirements_captioning_df.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
audioread==3.0.1
cachetools==5.5.0
cffi==1.17.1
click==8.1.8
contourpy==1.1.1
coverage==7.6.10
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
efficientnet_pytorch==0.7.1
einops==0.6.1
exceptiongroup==1.2.2
fastapi==0.115.8
fire==0.2.1
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==0.4.6
grpcio==1.69.0
h11==0.14.0
h5py==3.1.0
huggingface-hub==0.27.1
importlib_metadata==8.5.0
iniconfig==2.0.0
joblib==1.4.2
kaldiio==2.16.0a1
kiwisolver==1.4.7
librosa==0.9.0
llvmlite==0.39.1
Markdown==3.7
matplotlib==3.6.3
natsort==8.4.0
nltk==3.5
numba==0.56.4
oauthlib==3.2.2
pandas==1.1.3
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
protobuf==3.20.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pypeln==0.4.7
pytest==8.3.4
pytest-cov==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==5.1.2
regex==2024.11.6
requests-oauthlib==2.0.0
resampy==0.4.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.10.1
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.0
starlette==0.45.3
stopit==1.1.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torchlibrosa==0.1.0
transformers==4.47.1
typing_extensions==4.12.2
uvicorn==0.34.0
wandb==0.17.1
Werkzeug==3.1.3
zhon==1.1.5
zipp==3.21.0
